beautifying

+++ lily/custos.cc
@@ -54,7 +54,7 @@ Custos::print (SCM smob)
   Direction neutral_direction =
     to_dir (me->get_property ("neutral-direction"));
 
-  int pos = (int)Staff_symbol_referencer::get_rounded_position (me);
+  int pos = Staff_symbol_referencer::get_rounded_position (me);
   int sz = Staff_symbol_referencer::line_count (me)-1;
 
   String font_char = "custodes-" + style + "-";
